Break out ProduceBatch and ProduceBatchIterator
