+++ crates/fluvio-spu/src/smartengine/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod batch;
 pub(crate) mod file_batch;
+pub(crate) mod produce_batch;
 pub(crate) mod context;
 mod chain;
